Following better naming conventions
