+++ 07 transient reaction/OrderBooking.WebAPI/ApiRoutingConfiguration.cs
@@ -1,12 +1,12 @@
 using MessageHandler.EventSourcing.DomainModel;
-using OrderAggregate = OrderBooking.OrderBooking;
+using Booking = OrderBooking.OrderBooking;
 using OrderBooking.WebAPI.Controllers;
 using MessageHandler.EventSourcing.Projections;
 using OrderBooking.Projections;
 
 namespace OrderBooking.WebAPI;
 
-public static class MinimalApiConfig
+public static class ApiRoutingConfiguration
 {
     public static RouteGroupBuilder UseOrderBooking(
         this IEndpointRouteBuilder builder,
@@ -15,7 +15,7 @@ public static class MinimalApiConfig
         var orderBookings = groupBuilder(builder);
 
         orderBookings.MapPost("{id}",
-        async (IEventSourcedRepository<OrderAggregate> repo, string id, PlacePurchaseOrder cmd) =>
+        async (IEventSourcedRepository<Booking> repo, string id, PlacePurchaseOrder cmd) =>
         {
             var booking = await repo.Get(id);
             booking.PlacePurchaseOrder(cmd.PurchaseOrder, cmd.Name);
@@ -27,7 +27,7 @@ public static class MinimalApiConfig
         .Produces(StatusCodes.Status201Created);;
 
         orderBookings.MapGet("{id}", async(IRestoreProjections<Booking> projector, string id) =>
-            Results.Ok(await projector.Restore(nameof(OrderAggregate), id))
+            Results.Ok(await projector.Restore(nameof(Booking), id))
         ).Produces(StatusCodes.Status200OK);
 
         return orderBookings;
