+++ 10 manual task processing/OrderBooking.WebAPI/ApiRoutingConfiguration.cs
@@ -1,16 +1,12 @@
-using FastExpressionCompiler;
 using MessageHandler.EventSourcing.DomainModel;
-using Microsoft.Extensions.Azure;
-using OrderAggregate = OrderBooking.OrderBooking;
-using NotificationAggregate = NotificationPreferences.NotificationPreferences;
+using Booking = OrderBooking.OrderBooking;
 using OrderBooking.WebAPI.Controllers;
 using MessageHandler.EventSourcing.Projections;
-using OrderBooking.Projections;
 using Azure.Search.Documents;
 
 namespace OrderBooking.WebAPI;
 
-public static class MinimalApiConfig
+public static class ApiRoutingConfiguration
 {
     public static RouteGroupBuilder UseOrderBooking(
         this IEndpointRouteBuilder builder,
@@ -19,10 +15,10 @@ public static class MinimalApiConfig
         var orderBookings = groupBuilder(builder);
 
         orderBookings.MapPost("{id}",
-        async (IEventSourcedRepository<OrderAggregate> repo, string id, PlacePurchaseOrder cmd) =>
+        async (IEventSourcedRepository<Booking> repo, string id, PlacePurchaseOrder cmd) =>
         {
             var booking = await repo.Get(id);
-            booking.PlacePurchaseOrder(cmd.PurchaseOrder, cmd.BuyerId, cmd.Name);
+            booking.PlacePurchaseOrder(cmd.PurchaseOrder, cmd.Name);
 
             await repo.Flush();
 
@@ -31,7 +27,7 @@ public static class MinimalApiConfig
         .Produces(StatusCodes.Status201Created);;
 
         orderBookings.MapGet("{id}", async(IRestoreProjections<Booking> projector, string id) =>
-            Results.Ok(await projector.Restore(nameof(OrderAggregate), id))
+            Results.Ok(await projector.Restore(nameof(Booking), id))
         ).Produces(StatusCodes.Status200OK);
 
         orderBookings.MapGet("pending", async(SearchClient client) =>
@@ -44,7 +40,7 @@ public static class MinimalApiConfig
         .Produces(StatusCodes.Status200OK);
 
         orderBookings.MapPost("{bookingId}/confirm",
-        async (IEventSourcedRepository<OrderAggregate> repo, string bookingId) =>
+        async (IEventSourcedRepository<Booking> repo, string bookingId) =>
         {
             var aggregate = await repo.Get(bookingId);
             aggregate.ConfirmSalesOrder();
@@ -55,24 +51,5 @@ public static class MinimalApiConfig
         .Produces(StatusCodes.Status200OK);
 
         return orderBookings;
-    }
-    public static RouteGroupBuilder UseNotificationPreferences(
-        this IEndpointRouteBuilder builder,
-        Func<IEndpointRouteBuilder, RouteGroupBuilder> groupBuilder)
-    {
-        var notificationPreferences = groupBuilder(builder);
-
-        notificationPreferences.MapPost("{buyerId}",
-        async(IEventSourcedRepository<NotificationAggregate> repo, string buyerId, SetConfirmationMail command) =>
-        {
-            var aggregate = await repo.Get(buyerId);
-            aggregate.SetConfirmationEmail(command.EmailAddress);
-            await repo.Flush();
-
-            return Results.Ok(aggregate.Id);
-        })
-        .Produces(StatusCodes.Status200OK);;
-
-        return notificationPreferences;
     }
 }
