+++ 11 automated task processing/OrderBooking.WebAPI/ApiRoutingConfiguration.cs
@@ -1,13 +1,13 @@
 using MessageHandler.EventSourcing.DomainModel;
-using OrderAggregate = OrderBooking.OrderBooking;
+using Booking = OrderBooking.OrderBooking;
+using Preferences = NotificationPreferences.NotificationPreferences;
 using OrderBooking.WebAPI.Controllers;
 using MessageHandler.EventSourcing.Projections;
-using OrderBooking.Projections;
 using Azure.Search.Documents;
 
 namespace OrderBooking.WebAPI;
 
-public static class MinimalApiConfig
+public static class ApiRoutingConfiguration
 {
     public static RouteGroupBuilder UseOrderBooking(
         this IEndpointRouteBuilder builder,
@@ -16,10 +16,10 @@ public static class MinimalApiConfig
         var orderBookings = groupBuilder(builder);
 
         orderBookings.MapPost("{id}",
-        async (IEventSourcedRepository<OrderAggregate> repo, string id, PlacePurchaseOrder cmd) =>
+        async (IEventSourcedRepository<Booking> repo, string id, PlacePurchaseOrder cmd) =>
         {
             var booking = await repo.Get(id);
-            booking.PlacePurchaseOrder(cmd.PurchaseOrder, cmd.Name);
+            booking.PlacePurchaseOrder(cmd.PurchaseOrder, cmd.BuyerId, cmd.Name);
 
             await repo.Flush();
 
@@ -28,7 +28,7 @@ public static class MinimalApiConfig
         .Produces(StatusCodes.Status201Created);;
 
         orderBookings.MapGet("{id}", async(IRestoreProjections<Booking> projector, string id) =>
-            Results.Ok(await projector.Restore(nameof(OrderAggregate), id))
+            Results.Ok(await projector.Restore(nameof(Booking), id))
         ).Produces(StatusCodes.Status200OK);
 
         orderBookings.MapGet("pending", async(SearchClient client) =>
@@ -41,7 +41,7 @@ public static class MinimalApiConfig
         .Produces(StatusCodes.Status200OK);
 
         orderBookings.MapPost("{bookingId}/confirm",
-        async (IEventSourcedRepository<OrderAggregate> repo, string bookingId) =>
+        async (IEventSourcedRepository<Booking> repo, string bookingId) =>
         {
             var aggregate = await repo.Get(bookingId);
             aggregate.ConfirmSalesOrder();
@@ -52,5 +52,24 @@ public static class MinimalApiConfig
         .Produces(StatusCodes.Status200OK);
 
         return orderBookings;
+    }
+    public static RouteGroupBuilder UseNotificationPreferences(
+        this IEndpointRouteBuilder builder,
+        Func<IEndpointRouteBuilder, RouteGroupBuilder> groupBuilder)
+    {
+        var notificationPreferences = groupBuilder(builder);
+
+        notificationPreferences.MapPost("{buyerId}",
+        async(IEventSourcedRepository<Preferences> repo, string buyerId, SetConfirmationMail command) =>
+        {
+            var aggregate = await repo.Get(buyerId);
+            aggregate.SetConfirmationEmail(command.EmailAddress);
+            await repo.Flush();
+
+            return Results.Ok(aggregate.Id);
+        })
+        .Produces(StatusCodes.Status200OK);;
+
+        return notificationPreferences;
     }
 }
